requirements-mcp.txt
# WordOfPrompt MCP Dependencies
# Model Context Protocol support and related packages

# Core MCP packages
mcp>=1.0.0
anthropic-mcp>=0.1.0

# JSON-RPC and protocol support
jsonrpc-base>=2.1.0
jsonrpc-websocket>=3.1.4
jsonrpc-requests>=0.4.0

# WebSocket support for MCP transport
websockets>=12.0
websocket-client>=1.6.0

# Async support
aiofiles>=23.0.0
aiohttp>=3.9.0
asyncio-mqtt>=0.13.0

# Protocol serialization
msgpack>=1.0.0
cbor2>=5.4.0

# Enhanced JSON handling
orjson>=3.9.0
ujson>=5.8.0

# Validation and schema
jsonschema>=4.19.0
cerberus>=1.3.0

# Transport and communication
paramiko>=3.3.0  # SSH transport
pyserial>=3.5    # Serial transport

# Development and testing
pytest-asyncio>=0.21.0
aioresponses>=0.7.0
pytest-mock>=3.12.0

# Monitoring and observability
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
prometheus-client>=0.19.0

# Security
cryptography>=41.0.0
pynacl>=1.5.0

# Utilities
click>=8.1.0
rich>=13.0.0
typer>=0.9.0
